# Fix kernel_size != 3 on winograd arm/mali.