--- a/topi/tests/python/test_topi_conv2d_winograd.py
+++ b/topi/tests/python/test_topi_conv2d_winograd.py
@@ -83,7 +83,7 @@ def verify_conv2d_nchw(batch, in_channel, in_size, num_filter, kernel, stride, p
             func(a, w, c)
 
         rtol = 1e-5
-        if (kernel > 3):
+        if (kernel > 5):
           rtol = 2e-5
 
         tvm.testing.assert_allclose(c.asnumpy(), c_np, rtol=rtol)
@@ -110,16 +110,16 @@ def test_conv2d_nchw():
     with WinogradFallback():
 
         # inception v3 workloads
-        verify_conv2d_nchw(1, 128, 17, 192, 7, 1, 3, devices=['cuda'])
-        verify_conv2d_nchw(1, 128, 17, 128, 7, 1, 3, devices=['cuda'])
-        verify_conv2d_nchw(1, 160, 17, 160, 7, 1, 3, devices=['cuda'])
+        verify_conv2d_nchw(1, 128, 17, 192, 7, 1, 3)
+        verify_conv2d_nchw(1, 128, 17, 128, 7, 1, 3)
+        verify_conv2d_nchw(1, 160, 17, 160, 7, 1, 3)
 
         # resnet 18 workloads
         verify_conv2d_nchw(1, 64, 56, 64, 3, 1, 1)
         verify_conv2d_nchw(1, 128, 28, 128, 3, 1, 1)
         verify_conv2d_nchw(1, 256, 14, 256, 3, 1, 1)
         verify_conv2d_nchw(1, 512, 7, 512, 3, 1, 1)
-        verify_conv2d_nchw(1, 48,  35, 64, 5, 1, 2, devices=['cuda'])
+        verify_conv2d_nchw(1, 48, 35, 64, 5, 1, 2)
 
         # batch size = 2
         verify_conv2d_nchw(2, 64, 56, 64, 3, 1, 1)