@cbalint13 2019-07-28
> Is there a reason why we are disabling cuda tests for now?

* (kernel_size > 3) was running on cuda only.
* now with this runs on all targeted arch.

(kernel_size > 3) is new for our winograd, since its a larger kernel the accuracy may also drop, interesting that wasn't catched before (i think memoization data was changed).