@cbalint13 2019-07-28
Furthermore, 

* Also, in future would like to add ```float16``` too for this testcase, and for that rtol has to be even higher ~1e-4, so i think cannot avoid to handle different rtol values in our winograd conv2d tests.