@cbalint13 2019-07-29
> @merrymercy , @kevinthesun , @FrozenGene
> 
> Will update on [tophub](https://github.com/uwsampl/tvm-distro/tree/master/tophub) .
> Allow a little time for tophub update (need to fill `tile_size`), until then this PR is WiP.

It's done, review ready at https://github.com/uwsampl/tvm-distro/pull/8 .