@merrymercy 2019-07-31
Aha... I remember why I don't include `tile_size` as a tunable knob. This is not the `xgb` vs `xgb_knob` problem. It is because the search spaces for different tile sizes are different. But AutoTVM does not support this kind of space definition.

There are some knobs for tiling in the search space. However, they are dependent on tile size. For example, in the arm cpu's schedule, https://github.com/dmlc/tvm/blob/36702a7678573fa4ff9ea04c3920e60d590496c9/topi/python/topi/arm_cpu/conv2d.py#L390-L392. `P` is the number of winograd tiles. In one case, P is 196 if tile size is 4 while is 784 if tile size is 2. We construct the split knob according the the lengths of P, so different tile sizes will have different spaces.

Currently, AutoTVM does not support dependent relationship between tuning knobs. What it does is picking the first candidate of tile size (which is 2 according your construction) and initializing other knobs. In the above case it will enumerate all split schemes for P = 784. But some split schemes for P = 784 are invalid for P = 196, which will cause the AutoTVM hangs forever during feature extraction.

### Ideal solution
Support hierarchical space definition in AutoTVM

### Workaround
1. Fix the feature extraction. We hang on this line https://github.com/dmlc/tvm/blob/11da1ca31b8121e9a40c70f33ade27670d22e2d0/python/tvm/autotvm/tuner/xgboost_cost_model.py#L311 because we get exception when run this function https://github.com/dmlc/tvm/blob/11da1ca31b8121e9a40c70f33ade27670d22e2d0/python/tvm/autotvm/tuner/xgboost_cost_model.py#L328
2. Use `feature_type` = `knob` in the XGBTuner. I won't merge this solution to master.

You don't meet the problem when tuning for CUDA and mali because I added some restrictions to their schedule tempaltes and luckily they work around this problem.