@cbalint13 2019-07-31
@merrymercy ,

> Aha... I remember why I don't include `tile_size` as a tunable knob. This is not the `xgb` vs `xgb_knob` problem. It is because the search spaces for different tile sizes are different. But AutoTVM does not support this kind of space definition.

1. I was driven by the idea that too simple formula like (tile_size = 4 if H % 4 == 0 else 2) can be rather a simple knob with a small list. Now we afford it with flexible winograd matrix computation.
2.  As result fact, I posted in the header here, knob found better solution overriding mentioned static formula (2 vs 4 tile_size).

> 
> There are some knobs for tiling in the search space. However, they are dependent on tile size. For example, in the arm cpu's schedule,
> 
> https://github.com/dmlc/tvm/blob/36702a7678573fa4ff9ea04c3920e60d590496c9/topi/python/topi/arm_cpu/conv2d.py#L390-L392
> 
> . `P` is the number of winograd tiles. In one case, P is 196 if tile size is 4 while is 784 if tile size is 2. We construct the split knob according to the lengths of P, so different tile sizes will have different spaces.

* I also found this correlation and dependency to P in case of our arm/mali.

* Yes i agree, that is also my finding during tests regarding P. Would be great help me elaborate an exact relationship ```tile_size -> P``` . Otherwise, if we don't narrow down the possible lists,  we just can enumerate bulk ```range(0,256)``` for some of knob's there and let xgb find it out but it wold have to search in 10e+7 cfg.space than.

* Again, as fact, even ignoring this dependency for now we are still better than existing formula.
* In fact, for now, solutions with  this knob will be 2 or 4 (very rare 3 hits) in most cases.

> Currently, AutoTVM does not support dependent relationships between tuning knobs. What it does is picking the first candidate of tile size (which is 2 according to your construction) and initializing other knobs. In the above case, it will enumerate all split schemes for P = 784. But some split schemes for P = 784 are invalid for P = 196, which will cause the AutoTVM hangs forever during feature extraction.

* So, could help me find the relation of P given tile_size ?

> 
> ### Ideal solution
> Support hierarchical space definition in AutoTVM
> 
> ### Workaround
> 1. Fix the feature extraction. We hang on this line https://github.com/dmlc/tvm/blob/11da1ca31b8121e9a40c70f33ade27670d22e2d0/python/tvm/autotvm/tuner/xgboost_cost_model.py#L311
>     because we get exceptions when running this function https://github.com/dmlc/tvm/blob/11da1ca31b8121e9a40c70f33ade27670d22e2d0/python/tvm/autotvm/tuner/xgboost_cost_model.py#L328
> 2. Use `feature_type` = `knob` in the XGBTuner. I won't merge this solution to master.

* OK, i put xgb issue on higher pripority, it is trigerred 100% on such scenarios like #2111 

> 
> You don't meet the problem when tuning for CUDA and mali because I added some restrictions to their schedule templates and luckily they work around this problem. But for arm cpu we will meet this problem.

* Where are those restrictions exactly ?

> 
> Let me know if you want to try to fix it or want more information from me. Otherwise, I can fix it later.

Yes i give it a try to fix it, I'll update you on progress.

-----------

To sum up lets settle for now the order :

0. For now, as mentioned & tested,  our PR is still **better** than existing ```two cased conditional```. 
1. Try find xgb fea issue and repair it.
2. Lets see a smarter P given tile_size (please help me elaborate a generating formula).
3. If get ```tile_size->P``` we can try introduce hierarchical knob mode (i have an simple idea for that).


Arguing by point 0. , the PR would be mergeable, otherwise we transform this PR in a long multi PR.
Agree 1 +  2 + 3 are mandatory, i would help tackle each in upcoming another 3 PR (there will be too much debate if we keep in a single PR like this one).

What you think ?

----------