@cbalint13 2019-07-31
Opened a [disscuss topic](https://discuss.tvm.ai/t/winograd-variable-tile-size-knob/3565) and started first with ```xgb``` & ```hierarchical knob``` issue.