@merrymercy 2019-08-01
#3687 #3689 are necessary to merge this PR, otherwise this PR will make tuning on arm cpu fails.

I can merge this after the above two are merged.