@cbalint13 2019-08-01
> #3687 #3689 are necessary to merge this PR, otherwise this PR will make tuning on arm cpu fails.
> 
> I can merge this after the above two are merged.



> #3687 #3689 are necessary to merge this PR. Otherwise this PR will make tuning on arm cpu fails.
> 
> I can merge this after the above two are merged.

@merrymercy ,

May for  ```mali```expand other four vars in this PR , namely```yt```, ```t1```,``` t2``` and ```unroll``` [see here](https://discuss.tvm.ai/t/winograd-variable-tile-size-knob/3565/6?u=cbalint13) ?
Or lets do in other PR where I can show up with retuned/enhanced TOPI entries with comparing in numbers ?